// src/interfaces/IPaymentTranslate.ts
export interface IPaymentTranslate {
	paymentInformation: string;
	paymentByQRCode: string;
	download: string;
	warningPayment: string;
	or: string;
	copy: string;
	bank: string;
	bankingNumber: string;
	ownerAccount: string;
	tranferContent: string;
	timeLeft: string;
	cancel: string;
	confirmPayment: string;
	cancelOrChangeTicket: string;
	pay: string;
	waitingForConfirm: string;
	waitingDescription: string;
	backtoHome: string;
	saveTicket: string;
	informationCustomer: string;
	informationTrip: string;
	ticketCode: string;
	noteNoSupport: string;
	noteCheckBeforePayment: string;
	copied: string;
	totalPrice: string;
}